+++ mex/editor/ingest/state.py
@@ -11,7 +11,11 @@ from requests import HTTPError
 from mex.common.backend_api.connector import BackendApiConnector
 from mex.common.models import AnyExtractedModel, PaginatedItemsContainer
 from mex.editor.exceptions import escalate_error
-from mex.editor.ingest.models import IngestResult
+from mex.editor.ingest.models import (
+    ALL_AUX_PROVIDERS,
+    AuxProvider,
+    IngestResult,
+)
 from mex.editor.ingest.transform import transform_models_to_results
 from mex.editor.state import State
 from mex.editor.utils import resolve_editor_value
@@ -26,8 +30,8 @@ class IngestState(State):
     query_string: Annotated[str, Field(max_length=1000)] = ""
     current_page: Annotated[int, Field(ge=1)] = 1
     limit: Annotated[int, Field(ge=1, le=100)] = 50
-    current_aux_provider: str = "ldap"
-    aux_providers: list[str] = ["ldap", "orcid", "wikidata"]
+    current_aux_provider: AuxProvider = AuxProvider.LDAP
+    aux_providers: list[AuxProvider] = ALL_AUX_PROVIDERS
     is_loading: bool = True
 
     @rx.var(cache=False)
@@ -64,7 +68,7 @@ class IngestState(State):
         ].show_properties
 
     @rx.event
-    def set_current_aux_provider(self, value: str) -> None:
+    def set_current_aux_provider(self, value: AuxProvider) -> None:
         """Change the current aux provider."""
         self.current_aux_provider = value