+++ CHANGELOG.md
@@ -8,8 +8,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- infobox to ingest page for "ldap" and "wikidata"
+- new type for aux_provider instead of str
 
 ### Changes
+- type of aux provider from `str` to `AuxProvider`
 
 - prepare reflex update
 - de-duplicate pagination component